+++ apps/atlas/src/app/components/filters/RangeSliderFilter.tsx
