+++ apps/atlas/src/app/components/ui/BasePanel.tsx
