+++ apps/atlas/src/app/loot/LootPanel.tsx
@@ -1,20 +1,22 @@
+'use client';
+
 import { useQuery } from '@apollo/client';
 import { Tabs, Button } from '@bibliotheca-dao/ui-lib';
-import Danger from '@bibliotheca-dao/ui-lib/icons/danger.svg';
+import Bag from '@bibliotheca-dao/ui-lib/icons/bag.svg';
+
 import { useEffect, useMemo, useState } from 'react';
 import { useAccount as useL1Account } from 'wagmi';
-import { CryptFilter } from '@/components/filters/CryptFilter';
-import { CryptsOverviews } from '@/components/tables/CryptsOverviews';
-import { useCryptContext } from '@/context/CryptContext';
-import { getCryptsQuery } from '@/hooks/graphql/queries';
-import type { Crypt } from '@/types/index';
-import { BasePanel } from './BasePanel';
-import { PanelHeading } from './PanelComponents/PanelHeading';
+import { LootFilters } from '@/app/components/filters/LootFilters';
+import { BasePanel } from '@/app/components/ui/BasePanel';
+import { PanelHeading } from '@/app/components/ui/PanelHeading';
+import { LootOverviews } from '@/components/tables/LootOverviews';
+import { useLootContext } from '@/context/LootContext';
+import { getLootsQuery } from '@/hooks/graphql/queries';
+import type { Loot } from '@/types/index';
 
-export const CryptsPanel = () => {
-  // const { selectedId, openDetails, isDisplayLarge } = useAtlasContext();
+export const LootPanel = () => {
   const { address: l1Address } = useL1Account();
-  const { state, actions } = useCryptContext();
+  const { state, actions } = useLootContext();
 
   const limit = 50;
   const [page, setPage] = useState(1);
@@ -22,77 +24,64 @@ export const CryptsPanel = () => {
   const nextPage = () => setPage(page + 1);
 
   // Reset page on filter change. UseEffect doesn't do a deep compare
-  const filterDependencies = [
-    state.favouriteCrypt,
-    state.isLegendaryFilter,
+  const queryDependencies = [
+    state.favouriteLoot,
     state.searchIdFilter,
-    state.environmentsFilter,
-    state.statsFilter.numDoors,
-    state.statsFilter.numPoints,
-    state.statsFilter.size,
+    state.ratingFilter.bagGreatness,
+    state.ratingFilter.bagRating,
     state.selectedTab,
   ];
   useEffect(() => {
     setPage(1);
-  }, [...filterDependencies]);
+  }, [...queryDependencies]);
 
-  const isCryptPanel = true;
-  const tabs = ['Your Crypts', 'All Crypts', 'Favourite Crypts'];
+  const isLootPanel = true;
+  const tabs = ['Your Loot', 'All Loot', 'Favourite Loot'];
 
   const variables = useMemo(() => {
     const where: any = {};
     if (state.searchIdFilter) {
       where.id = state.searchIdFilter;
     } else if (state.selectedTab === 2) {
-      where.id_in = [...state.favouriteCrypt];
+      where.id_in = [...state.favouriteLoot];
     }
 
     if (state.selectedTab === 0) {
       where.currentOwner = l1Address?.toLowerCase();
     }
-
-    if (state.isLegendaryFilter) {
-      where.name_starts_with = "'";
-    }
-
-    where.numDoors_gte = state.statsFilter.numDoors.min;
-    where.numDoors_lte = state.statsFilter.numDoors.max;
-    where.numPoints_gte = state.statsFilter.numPoints.min;
-    where.numPoints_lte = state.statsFilter.numPoints.max;
-    where.size_gte = state.statsFilter.size.min;
-    where.size_lte = state.statsFilter.size.max;
-
-    if (state.environmentsFilter.length > 0) {
-      where.environment_in = [...state.environmentsFilter];
-    }
+    where.bagGreatness_gte = state.ratingFilter.bagGreatness.min;
+    where.bagGreatness_lte = state.ratingFilter.bagGreatness.max;
+    where.bagRating_gte = state.ratingFilter.bagRating.min;
+    where.bagRating_lte = state.ratingFilter.bagRating.max;
 
     return {
       first: limit,
       skip: limit * (page - 1),
       where,
+      orderBy: 'minted',
+      orderDirection: 'asc',
     };
-  }, [l1Address, page, ...filterDependencies]);
+  }, [l1Address, page, ...queryDependencies]);
 
-  const { loading, data } = useQuery<{
-    dungeons: Crypt[];
-  }>(getCryptsQuery, {
+  const { loading, data } = useQuery<{ bags: Loot[] }>(getLootsQuery, {
     variables,
-    skip: !isCryptPanel,
+    skip: !isLootPanel,
   });
 
   const showPagination = () =>
     state.selectedTab === 1 &&
-    (page > 1 || (data?.dungeons?.length ?? 0) === limit);
+    (page > 1 || (data?.bags?.length ?? 0) === limit);
 
-  const hasNoResults = () => !loading && (data?.dungeons?.length ?? 0) === 0;
+  const hasNoResults = () => !loading && (data?.bags?.length ?? 0) === 0;
 
   return (
-    <BasePanel open={isCryptPanel}>
+    <BasePanel open={isLootPanel}>
       <PanelHeading
-        heading="Crypts"
+        heading="Loot"
         action={actions.updateSearchIdFilter}
         searchFilterValue={state.searchIdFilter}
       />
+
       <Tabs
         selectedIndex={state.selectedTab}
         onChange={actions.updateSelectedTab as any}
@@ -104,14 +93,14 @@ export const CryptsPanel = () => {
         </Tabs.List>
       </Tabs>
       <div>
-        <CryptFilter />
+        <LootFilters />
         {loading && (
           <div className="flex flex-col items-center w-20 gap-2 mx-auto my-40 animate-pulse">
-            <Danger className="block w-20 fill-current" />
+            <Bag className="block w-20 fill-current" />
             <h2>Loading</h2>
           </div>
         )}
-        <CryptsOverviews dungeons={data?.dungeons ?? []} />
+        <LootOverviews bags={data?.bags ?? []} />
       </div>
 
       {hasNoResults() && (
@@ -129,7 +118,7 @@ export const CryptsPanel = () => {
                 className="whitespace-nowrap"
                 onClick={() => actions.updateSelectedTab(1)}
               >
-                See All Crypts
+                See All Loot
               </Button>
             )}
           </div>
@@ -141,10 +130,7 @@ export const CryptsPanel = () => {
           <Button onClick={previousPage} disabled={page === 1}>
             Previous
           </Button>
-          <Button
-            onClick={nextPage}
-            disabled={data?.dungeons?.length !== limit}
-          >
+          <Button onClick={nextPage} disabled={data?.bags?.length !== limit}>
             Next
           </Button>
         </div>
