+++ apps/atlas/src/app/realm/(realmList)/page.tsx
@@ -1,6 +1,6 @@
 // Import your Client Component
+import { BasePanel } from '@/app/components/ui/BasePanel';
 import { ArtBackground } from '@/components/map/ArtBackground';
-import { BasePanel } from '@/components/panels/BasePanel';
 import { RealmProvider } from '@/context/RealmContext';
 import { getRealms } from '@/lib/realm/getRealms';
 import { RealmListPanel } from './RealmListPanel';
