+++ apps/atlas/src/components/panels/Account/MyLoot.tsx
@@ -3,7 +3,7 @@ import { Button } from '@bibliotheca-dao/ui-lib';
 import Bag from '@bibliotheca-dao/ui-lib/icons/bag.svg';
 import { useEffect, useMemo, useState } from 'react';
 import { useAccount as useL1Account } from 'wagmi';
-import { LootFilters } from '@/components/filters/LootFilters';
+import { LootFilters } from '@/app/components/filters/LootFilters';
 import { LootOverviews } from '@/components/tables/LootOverviews';
 import { useLootContext } from '@/context/LootContext';
 import { getLootsQuery } from '@/hooks/graphql/queries';
