+++ apps/atlas/src/components/panels/Account/MyRealms.tsx
@@ -8,8 +8,8 @@ import { useRouter, usePathname, useSearchParams } from 'next/navigation';
 import { useEffect, useMemo, useState } from 'react';
 import { useAccount as useL1Account } from 'wagmi';
 import { RealmOverviews } from '@/app/realm/(realmList)/RealmOverviews';
-import { RealmsFilter } from '@/components/filters/RealmsFilter';
-import { SearchFilter } from '@/components/filters/SearchFilter';
+import { RealmsFilter } from '@/app/components/filters/RealmsFilter';
+import { SearchFilter } from '@/app/components/filters/SearchFilter';
 import { RealmsMax } from '@/constants/index';
 import { useRealmContext } from '@/context/RealmContext';
 import type { RealmTraitType } from '@/generated/graphql';
