+++ apps/atlas/src/context/CryptContext.tsx
@@ -1,3 +1,4 @@
+'use client';
 import type { Dispatch } from 'react';
 import { createContext, useContext, useReducer } from 'react';
 import { storage } from '@/util/localStorage';
