+++ apps/atlas/src/context/GaContext.tsx
@@ -1,3 +1,4 @@
+'use client';
 import type { Dispatch } from 'react';
 import { createContext, useContext, useReducer } from 'react';
 import type { OrderType } from '@/generated/graphql';
